% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narmviz.R
\name{draw_rectangle}
\alias{draw_rectangle}
\title{Create a Rectangle Data Frame}
\usage{
draw_rectangle(x, y, w, h)
}
\arguments{
\item{x}{Numeric, x-coordinate of the lower-left corner.}

\item{y}{Numeric, y-coordinate of the lower-left corner.}

\item{w}{Numeric, width of the rectangle.}

\item{h}{Numeric, height of the rectangle.}
}
\value{
A data frame with x and y coordinates defining the rectangle.
}
\description{
Create a Rectangle Data Frame
}
