% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narmviz.R
\name{visualize}
\alias{visualize}
\title{Visualize a Rule on a Dataset}
\usage{
visualize(
  rule,
  dataset,
  path = NULL,
  allfeatures = FALSE,
  antecedent = TRUE,
  consequent = TRUE,
  timeseries = FALSE,
  intervalcolumn = "interval",
  interval = 0
)
}
\arguments{
\item{rule}{List containing antecedent and consequent attributes.}

\item{dataset}{Data frame containing the transaction data.}

\item{path}{Character, file path to save the visualization (optional).}

\item{allfeatures}{Logical, whether to visualize all features.}

\item{antecedent}{Logical, whether to visualize antecedent attributes.}

\item{consequent}{Logical, whether to visualize consequent attributes.}

\item{timeseries}{Logical, whether the dataset is a time series.}

\item{intervalcolumn}{Character, name of the column representing intervals.}

\item{interval}{Integer, specific interval to visualize.}
}
\value{
A combined plot displaying the rule visualization.
}
\description{
Visualize a Rule on a Dataset
}
